+++ packages/code-studio/src/main/AppMainContainer.test.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import React from 'react';
 import { Provider } from 'react-redux';
-import { render, screen } from '@testing-library/react';
+import { act, render, screen } from '@testing-library/react';
 import { ConnectionContext } from '@deephaven/app-utils';
 import { ToolType } from '@deephaven/dashboard-core-plugins';
 import {
@@ -112,17 +112,39 @@ function renderAppMainContainer({
     </Provider>
   );
 }
+
+const EMPTY_LAYOUT = {
+  filterSets: [],
+  layoutConfig: [],
+  links: [],
+  version: 2,
+};
+
 let mockProp = {};
 let mockId = DEFAULT_DASHBOARD_ID;
+let mockIteration = 0;
 jest.mock('@deephaven/dashboard', () => ({
   ...jest.requireActual('@deephaven/dashboard'),
   __esModule: true,
   LazyDashboard: jest.fn(({ hydrate }) => {
+    const { useMemo } = jest.requireActual('react');
+    // We use the `key` to determine how many times this LazyDashboard component was re-rendered with a new key
+    // When rendered with a new key, the `useMemo` will be useless and will return a new key
+    const key = useMemo(() => {
+      const newKey = `${mockIteration}`;
+      mockIteration += 1;
+      return newKey;
+    }, []);
     const result = hydrate(mockProp, mockId);
     if (result.fetch != null) {
       result.fetch();
     }
-    return <p>{JSON.stringify(result)}</p>;
+    return (
+      <>
+        <p>{JSON.stringify(result)}</p>
+        <p data-testid="dashboard-key">{key}</p>
+      </>
+    );
   }),
   default: jest.fn(),
 }));
@@ -136,6 +158,8 @@ beforeEach(() => {
     cb(0);
     return 0;
   });
+  mockProp = {};
+  mockIteration = 0;
 });
 
 afterEach(() => {
@@ -239,5 +263,38 @@ describe('hydrates widgets correctly', () => {
       )
     ).toBeTruthy();
     expect(objectFetcher).toHaveBeenCalled();
+  });
+});
+
+describe('imports layout correctly', () => {
+  it('uses a new key when layout is imported', async () => {
+    renderAppMainContainer();
+
+    expect(screen.getByText('{"localDashboardId":"default"}')).toBeTruthy();
+
+    const oldKey = screen.getByTestId('dashboard-key').textContent ?? '';
+    expect(oldKey.length).not.toBe(0);
+
+    await act(async () => {
+      const text = JSON.stringify(EMPTY_LAYOUT);
+      const file = TestUtils.createMockProxy<File>({
+        text: () => Promise.resolve(text),
+        name: 'layout.json',
+        type: 'application/json',
+      });
+
+      // Technically, the "Import Layout" button in the panels list is what the user clicks on to show the file picker
+      // However, the testing library uses the `.upload` command on the `input` element directly, which we don't display
+      // So just fetch it by testid and use the `.upload` command: https://testing-library.com/docs/user-event/utility/#upload
+      const importInput = screen.getByTestId('input-import-layout');
+      await userEvent.upload(importInput, file);
+    });
+
+    expect(screen.getByText('{"localDashboardId":"default"}')).toBeTruthy();
+
+    const newKey = screen.getByTestId('dashboard-key').textContent ?? '';
+
+    expect(newKey.length).not.toBe(0);
+    expect(newKey).not.toBe(oldKey);
   });
 });