+++ packages/components/src/spectrum/navigation.ts
@@ -1,0 +1,12 @@
+export {
+  Breadcrumbs,
+  type SpectrumBreadcrumbsProps as BreadcrumbsProps,
+  Link,
+  type SpectrumLinkProps as LinkProps,
+  TabList,
+  type SpectrumTabListProps as TabListProps,
+  TabPanels,
+  type SpectrumTabPanelsProps as TabPanelsProps,
+  Tabs,
+  type SpectrumTabsProps as TabsProps,
+} from '@adobe/react-spectrum';