+++ packages/components/src/spectrum/overlays.ts
@@ -1,0 +1,12 @@
+export {
+  AlertDialog,
+  type SpectrumAlertDialogProps as AlertDialogProps,
+  ContextualHelp,
+  type SpectrumContextualHelpProps as ContextualHelpProps,
+  Dialog,
+  type SpectrumDialogProps as DialogProps,
+  DialogContainer,
+  type SpectrumDialogContainerProps as DialogContainerProps,
+  DialogTrigger,
+  type SpectrumDialogTriggerProps as DialogTriggerProps,
+} from '@adobe/react-spectrum';